[ELY-1711] Move the HTTP authentication framework into the the 'wildfly-elytron-http' module.
