+++ wildfly-elytron-http/pom.xml
@@ -37,7 +37,7 @@
     <dependencies>
         <dependency>
             <groupId>org.wildfly.security</groupId>
-            <artifactId>wildfly-elytron-titan</artifactId>
+            <artifactId>wildfly-elytron-auth-server</artifactId>
         </dependency>
         
         <dependency>
