+++ wildfly-elytron-http/src/main/java/org/wildfly/security/http/HttpConstants.java
