+++ wildfly-elytron-titan/pom.xml
@@ -345,6 +345,10 @@
         </dependency>
         <dependency>
             <groupId>org.wildfly.security</groupId>
+            <artifactId>wildfly-elytron-http</artifactId>
+        </dependency>        
+        <dependency>
+            <groupId>org.wildfly.security</groupId>
             <artifactId>wildfly-elytron-permission</artifactId>
         </dependency>
         <dependency>
